requirements.txt
python-dotenv
# Optional dependencies
# Pillow
# cssmin
# jsmin
